refactor: move every component to its own file
