+++ src/components/About.tsx
@@ -1,0 +1,40 @@
+
+function About() {
+  return (
+    <div className="container">
+      <div className="landingAbout">
+        <p>002</p>
+        <p>About</p>
+        <span className="bio">
+          Santiago Ocampo is a Professor of Fine Arts and Programmer,
+          specialized in drawing and self-taught in the use of technology. My
+          work involves the use of analogue knowledge and the use of new media
+          to express my vision through digital pieces, in which the use of
+          color, lighting and animation are combined.
+        </span>
+        <h1 className="bioEnjoy">
+          I ENJOY: GAMING, MUSIC, RESEARCHING, LEARNING NEW SKILLS, 3D PRINTING
+        </h1>
+        {/* <iframe
+          src="https://my.spline.design/untitled-e3fbad4906352c5a176c4b0e2913680e/"
+          frameborder="0"
+          width="80%"
+          height="30%"
+          title="ProfilePic"
+        ></iframe> */}
+        <div className="tools">
+          <p>Software & Tools:</p>
+          <p>Blender 4.X</p>
+          <p>Unreal Engine 5.X</p>
+          <p>Touchdesigner</p>
+          <p>Cables.GL</p>
+          <p>Substance 3D Modeler</p>
+          <p>Figma</p>
+          <p>Resolume Arena</p>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default About;
