+++ src/components/Contact.tsx
@@ -1,0 +1,26 @@
+function Contact() {
+  return (
+    <div className="container">
+      <div className="landingContact">
+        <p>003</p>
+        <p>Contact</p>
+        <div className="message">
+          <p>Want to work together?</p>
+          <p>Send me a message</p>
+          <div className="mail">
+            <h1>SANTIAGO.OCAMPOMAIL@GMAIL.COM</h1>
+          </div>
+          <div className="contactFooter">
+            <p>Santiago Ocampo 3D Generalist</p>
+            <p>Instagram, Behance</p>
+            <p className="designedBy">
+              Design & Developmnet by Santiago Ocampo
+            </p>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Contact;
